rename controller interface -> h6x controller interface
