+++ .github/workflows/ci_humble.yml
@@ -33,4 +33,5 @@ jobs:
           import-token: ${{ secrets.H6X_REPO_TOKEN }}
           vcs-repo-file-url: build_depends.repos
           package-name: |
+            h6x_controller_interface
             image_sensor_broadcaster
