+++ .vscode/c_cpp_properties.json
@@ -10,7 +10,7 @@
         "/opt/ros/humble/src/gmock_vendor/include/**",
         "/usr/include/**",
         "${workspaceFolder}/image_sensor_broadcaster/include/**",
-        "${workspaceFolder}/controller_interface/include/**",
+        "${workspaceFolder}/h6x_controller_interface/include/**",
         "${workspaceFolder}/../../install/image_sensor_broadcaster/include/**",
       ],
       "name": "ROS",
