+++ h6x_controller_interface/include/semantic_components/image_sensor.hpp
