+++ image_sensor_broadcaster/package.xml
@@ -10,7 +10,7 @@
   <buildtool_depend>ament_cmake_auto</buildtool_depend>
 
   <depend>backward_ros</depend>
-  <depend>controller_interface</depend>
+  <depend>h6x_controller_interface</depend>
   <depend>generate_parameter_library</depend>
   <depend>hardware_interface</depend>
   <depend>pluginlib</depend>
